tg_bot/requirements_tg.txt
aio-pika==9.5.5
python-telegram-bot==22.1
nest-asyncio==1.6.0
